requirements_test.txt
coverage==4.2
mock>=1.0.1
flake8>=2.1.0
tox>=1.7.0
codecov
django-compat>=1.0.8
djangorestframework>=3.2
django-braces>=1.10.0
